README: Update bootstrap example to use 'build' directory under source
When the example was first written, developers commonly placed their
build trees as siblings of the source tree.  Nowadays many developers
use a `build/` directory under the top-level source directory instead.
Update the example with the modern convention to avoid requiring the
reader to understand that `cmake-source` is a placeholder for the path
to the source tree.