Ninja: Revert exact collation dependencies for 3.27
Revert commit b6a5382217 (Ninja: depend on language module information
files directly, 2023-02-10, v3.27.0-rc1~502^2) from !8197.

This reverts the "exact dependencies" for collation inputs and returns
to "get all targets" and target-ordering. Use of exact dependencies
caused a parade of use cases that had not been tested previously to be
found and need fixing over the 3.27 release series. To stop the flow on
3.27, revert to the 3.26 strategy.  We will continue in 3.28.

Note that this is a restoration of 3.26 semantics where incremental
rebuilds may be subtly incorrect in the presence of stale
`<LANG>Modules.json` files. However, since C++ support is experimental
and Fortran has always had this problem as of 3.27, it is not considered
a regression.

See: #25112
See: #25123
See: #25252
See: #25365
See: #25417
See: #25425