Help: Fix ctest(1) manual links to www.cdash.org
When the link was updated to `https` by commit 52eac4573d (Help: Fix
link to cdash.org from CTest manual, 2021-04-21, v3.21.0-rc1~262^2~2)
the markup was incorrectly adjusted to show the link as part of the
"See Also" section.  It is meant to be the link destination for links
in prose elsewhere in the manual.  Fix the markup and move it to a
clearer location.  Also update the link to resolve a redirect.