Tests: Avoid procedure stack overflow in broken libc on Elbrus
On E2K architecture, there is at least one known version of libc
that is built with a bug that leads to a kernel error like
`procedure stack could not be copied` visible in `dmesg` in several
tests (and such test immediately receives `SIGSEGV` and fails).
It was first detected in !8665, and after a long investigation,
the culprit was finally found (MCST bugzilla internal bug 124224).

Avoid running tests known to fail if such a version of libc is detected.